повторная защита 6
